% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{apply_over}
\alias{apply_over}
\title{Applies a function over chains}
\usage{
apply_over(f, data)
}
\arguments{
\item{f}{Function to apply}

\item{data}{A single data.frame or list of data.frames}
}
\description{
Applies a function over single chain or list of chains
}

