% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{density_plot}
\alias{density_plot}
\title{Density distributions of specified parameters.}
\usage{
density_plot(data, pars = NULL, xlab = "standardized value",
  scales = "free", ...)
}
\arguments{
\item{data}{A single data.frame or list of data.frames}

\item{pars}{List of parameters to include}

\item{xlab}{label for xaxis}

\item{scales}{By default set to "free", indicating each subplot will have different
scale on yaxis. Set to "fixed" to enforce similar axes over all plots.}

\item{...}{Extra augments for \code{ggplot2::theme} function.}
}
\description{
Density distributions of specified parameters.
}

