% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{drop_columns}
\alias{drop_columns}
\title{Drop specified columns from chains}
\usage{
drop_columns(data, to_drop)
}
\arguments{
\item{data}{A single data.frame or list of data.frames}

\item{to_drop}{vector of columns names to remove}
}
\description{
Drop specified columns from chains
}

