% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{gather_samples}
\alias{gather_samples}
\title{Gather columns into key-value pairs.}
\usage{
gather_samples(data, pars = NULL, drop = stateline_sample_names(),
  thin = NULL)
}
\arguments{
\item{data}{A single data.frame or list of data.frames}

\item{pars}{List of parameters to include}

\item{drop}{List of columns to drop. By default stateline's extra columns
(energy, sigma, beta, accepted, swap_type) are removed.}

\item{thin}{If set to a numeric value, will thin samples down to
specified length using \code{thin\_chains} function.}
}
\value{
data.frame with columns stack, iteration, parameter, value
}
\description{
Gather_samples is built off \code{tidyr}'s \code{gather} function, which
takes multiple columns and collapses into key-value pairs with other
columns being "stack", "iteration".
}

