% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{install_scripts}
\alias{install_scripts}
\title{Install script}
\usage{
install_scripts(path)
}
\arguments{
\item{path}{Location to install the script}
}
\description{
Install script into the path
}

