% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{is.stack}
\alias{is.stack}
\title{Check if object is a data.frame or a list of data.frames}
\usage{
is.stack(data)
}
\arguments{
\item{data}{A single data.frame or list of data.frames}
}
\description{
Check if object is a data.frame or a list of data.frames
}

