% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_chain}
\alias{load_chain}
\title{Load a single stateline output file}
\usage{
load_chain(stack, path = "output", sample_names = NULL)
}
\arguments{
\item{stack}{Number of chain}

\item{path}{Directory where samples are saved}

\item{sample_names}{Vector of names for samples being fitted in the model}
}
\value{
Returns a data.frame with named columns.
}
\description{
Load a single stateline output file.
We use the \code{read\_csv} function from the \code{readr}
package, resulting in faster load times and better display.
}

