% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_chains}
\alias{load_chains}
\title{Load a collection of stateline output files}
\usage{
load_chains(path, nstacks = 4, merge = FALSE, ...)
}
\arguments{
\item{path}{Directory where samples are saved}

\item{nstacks}{Number of stacks to load}

\item{merge}{Boolean indicating whether stacks should be merged into
single data.frame}

\item{...}{Arguments to \code{chain}. Including a vector
\code{sample\_names}, a vector of names for samples being fitted in the model
will give data with appropriate column names.}
}
\value{
Returns a list with elements \code{stack\_1, stack\_2},
each being a data.frame with named columns, or in the case where
\code{merge=TRUE} a singel merged data.frame
}
\description{
Load a collection of stateline output files
}

