% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{merge_chains}
\alias{merge_chains}
\title{Create single unified data.frame from a list of data.frames}
\usage{
merge_chains(x)
}
\arguments{
\item{x}{A list, or something coercible to a list}
}
\description{
Create a unified data.frame from a list by rbinding files together
}

