% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{remove_warmup}
\alias{remove_warmup}
\title{Remove samples from defined warmup period}
\usage{
remove_warmup(data, warmup)
}
\arguments{
\item{data}{Merged or unmerged list of samples from stateline output,
obatined by running \code{load_chains}}

\item{warmup}{Length of warmup period (number of samples)}
}
\description{
Culls all samples taken prior to specified point \code{warmup}
from chain or of chains.
}

