% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\name{stateline_server}
\alias{stateline_server}
\title{Start a stateline server}
\usage{
stateline_server(config, output = "output", name = "stateline_server",
  rm = TRUE)
}
\arguments{
\item{config}{Configuration file}

\item{output}{Location to store output on the host computer}

\item{name}{Name of the server container}

\item{rm}{Delete the container on exit (default is to delete)}
}
\description{
Start a stateline server in a docker container
}

