% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{stateline_theme}
\alias{stateline_theme}
\title{Default theme used in plotting stateline objects}
\usage{
stateline_theme(...)
}
\arguments{
\item{...}{Extra augments for \code{ggplot2::theme} function.}
}
\description{
Default theme used in plotting stateline objects (via \code{ggplot2})
}

