% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stateliner.R
\name{stateliner}
\alias{stateliner}
\title{Stateline R interface}
\usage{
stateliner(target, config, address)
}
\arguments{
\item{target}{Target function; must be a function taking a single
argument \code{x} and returning a \emph{negative} log-likelhood.}

\item{config}{Path to the configuration file}

\item{address}{Address of the delegator}
}
\description{
Run stateline to sample from the given target function.
}

