% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{thin_chains}
\alias{thin_chains}
\title{Thin chains down to specified number of samples}
\usage{
thin_chains(data, length.out = 2000)
}
\arguments{
\item{data}{Chain or list of chains to be thinned}

\item{length.out}{Desired number of samples in output}
}
\value{
data.frame of similar structure to input, but with fewer
number of rows.
}
\description{
Thin chains down to specified number of samples.
Takes samples at fixed intervals across available length.
}

