% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{trace_plot}
\alias{trace_plot}
\title{Trace plot of specified parameters against iteration number.}
\usage{
trace_plot(data, thin = 2000, pars = NULL, xlab = "iterations",
  ylab = "standardized value", scales = "free", ...)
}
\arguments{
\item{data}{A single data.frame or list of data.frames}

\item{thin}{Numeric value (default =2000) indicating number of samples to
include in plot. Chains are thinned to specified length using \code{thin\_chains}.}

\item{pars}{List of parameters to include}

\item{xlab}{label for xaxis}

\item{ylab}{label for yaxis}

\item{scales}{By default set to "free", indicating each subplot will have different
scale on yaxis. Set to "fixed" to enforce similar axes over all plots.}

\item{...}{Extra augments for \code{ggplot2::theme} function.}
}
\description{
Trace plot of specified parameters against iteration number.
Good sampling is indicated by convergence and nice mixing of chains.
}

